test, starknet_scripts: initializing darkpool w/ verifier in integration tests
