+++ tests/src/utils.rs
@@ -2,9 +2,13 @@ use ark_ff::{BigInteger, PrimeField};
 use dojo_test_utils::sequencer::{Environment, StarknetConfig, TestSequencer};
 use eyre::{eyre, Result};
 use katana_core::{constants::DEFAULT_INVOKE_MAX_STEPS, sequencer::SequencerConfig};
+use merlin::HashChainTranscript;
 use mpc_bulletproof::{
-    r1cs::{R1CSProof, SparseReducedMatrix, SparseWeightRow},
-    InnerProductProof,
+    r1cs::{
+        CircuitWeights, ConstraintSystem, Prover, R1CSProof, SparseReducedMatrix, SparseWeightRow,
+        Variable, Verifier,
+    },
+    BulletproofGens, InnerProductProof, PedersenGens,
 };
 use mpc_stark::algebra::{scalar::Scalar, stark_curve::StarkPoint};
 use num_bigint::{BigUint, RandBigInt};
@@ -165,31 +169,6 @@ pub fn insert_scalar_to_ark_merkle_tree(
         .map_err(|e| eyre!("Error updating arkworks merkle tree: {}", e))?;
 
     Ok(Scalar::from_be_bytes_mod_order(&ark_merkle_tree.root()))
-}
-
-pub fn get_dummy_proof() -> R1CSProof {
-    R1CSProof {
-        A_I1: StarkPoint::generator(),
-        A_O1: StarkPoint::generator(),
-        S1: StarkPoint::generator(),
-        A_I2: StarkPoint::generator(),
-        A_O2: StarkPoint::generator(),
-        S2: StarkPoint::generator(),
-        T_1: StarkPoint::generator(),
-        T_3: StarkPoint::generator(),
-        T_4: StarkPoint::generator(),
-        T_5: StarkPoint::generator(),
-        T_6: StarkPoint::generator(),
-        t_x: Scalar::random(&mut thread_rng()),
-        t_x_blinding: Scalar::random(&mut thread_rng()),
-        e_blinding: Scalar::random(&mut thread_rng()),
-        ipp_proof: InnerProductProof {
-            L_vec: vec![],
-            R_vec: vec![],
-            a: Scalar::random(&mut thread_rng()),
-            b: Scalar::random(&mut thread_rng()),
-        },
-    }
 }
 
 // --------------------------
@@ -435,5 +414,173 @@ impl CalldataSerializable for CircuitParams {
             )
             .chain(self.c.to_calldata().into_iter())
             .collect()
+    }
+}
+
+// -------------------------
+// | DUMMY CIRCUIT HELPERS |
+// -------------------------
+
+// The dummy circuit we're using is a very simple circuit with 4 witness elements,
+// 3 multiplication gates, and 2 linear constraints. In total, it is parameterized as follows:
+// n = 3
+// n_plus = 4
+// k = log2(n_plus) = 2
+// m = 4
+// q = 8 (The total number of linear constraints is 8 because there are 3 multiplication gates, and 2 linear constraints per multiplication gate)
+
+// The circuit is defined as follows:
+//
+// Witness:
+// a, b, x, y (all scalars)
+//
+// Circuit:
+// m_1 = multiply(a, b)
+// m_2 = multiply(x, y)
+// m_3 = multiply(m_1, m_2)
+// constrain(a - 69)
+// constrain(m_3 - 420)
+
+// Bearing the following weights:
+// W_L = [[(0, -1)], [], [(1, -1)], [], [(2, -1)], [], [], []]
+// W_R = [[], [(0, -1)], [], [(1, -1)], [], [(2, -1)], [], []]
+// W_O = [[], [], [], [], [(0, 1)], [(1, 1)], [], [(2, 1)]]
+// W_V = [[(0, -1)], [(1, -1)], [(2, -1)], [(3, -1)], [], [], [(0, -1)], []]
+// c = [(6, 69), (7, 420)]
+
+pub const DUMMY_CIRCUIT_N: usize = 3;
+pub const DUMMY_CIRCUIT_N_PLUS: usize = 4;
+pub const DUMMY_CIRCUIT_K: usize = 2;
+pub const DUMMY_CIRCUIT_M: usize = 4;
+pub const DUMMY_CIRCUIT_Q: usize = 8;
+
+pub fn singleprover_prove_dummy_circuit() -> Result<(R1CSProof, Vec<StarkPoint>)> {
+    debug!("Generating proof for dummy circuit...");
+    let mut transcript = HashChainTranscript::new(TRANSCRIPT_SEED.as_bytes());
+    let pc_gens = PedersenGens::default();
+    let prover = Prover::new(&pc_gens, &mut transcript);
+
+    let witness = get_dummy_circuit_witness();
+
+    prove(prover, witness)
+}
+
+fn get_dummy_circuit_witness() -> Vec<Scalar> {
+    let a = Scalar::from(69);
+    let b = Scalar::from(420) * a.inverse();
+    let x = Scalar::one();
+    let y = Scalar::one();
+    vec![a, b, x, y]
+}
+
+fn prove(mut prover: Prover, witness: Vec<Scalar>) -> Result<(R1CSProof, Vec<StarkPoint>)> {
+    let mut rng = thread_rng();
+
+    // Commit to the witness
+    let a_blind = Scalar::random(&mut rng);
+    let b_blind = Scalar::random(&mut rng);
+    let x_blind = Scalar::random(&mut rng);
+    let y_blind = Scalar::random(&mut rng);
+
+    let (a_comm, a_var) = prover.commit(witness[0], a_blind);
+    let (b_comm, b_var) = prover.commit(witness[1], b_blind);
+    let (x_comm, x_var) = prover.commit(witness[2], x_blind);
+    let (y_comm, y_var) = prover.commit(witness[3], y_blind);
+
+    // Apply the constraints
+    apply_dummy_circuit_constraints(a_var, b_var, x_var, y_var, &mut prover);
+
+    // Generate the proof
+    let bp_gens = BulletproofGens::new(8 /* gens_capacity */, 1 /* party_capacity */);
+    let proof = prover
+        .prove(&bp_gens)
+        .map_err(|e| eyre!("error generating proof: {e}"))?;
+
+    Ok((proof, vec![a_comm, b_comm, x_comm, y_comm]))
+}
+
+fn apply_dummy_circuit_constraints<CS: ConstraintSystem>(
+    a: Variable,
+    b: Variable,
+    x: Variable,
+    y: Variable,
+    cs: &mut CS,
+) {
+    let (_, _, m_1) = cs.multiply(a.into(), b.into());
+    let (_, _, m_2) = cs.multiply(x.into(), y.into());
+    let (_, _, m_3) = cs.multiply(m_1.into(), m_2.into());
+    cs.constrain(a - Scalar::from(69));
+    cs.constrain(m_3 - Scalar::from(420));
+}
+
+pub fn prep_dummy_circuit_verifier(verifier: &mut Verifier, witness_commitments: Vec<StarkPoint>) {
+    // Allocate witness commitments into circuit
+    let a_var = verifier.commit(witness_commitments[0]);
+    let b_var = verifier.commit(witness_commitments[1]);
+    let x_var = verifier.commit(witness_commitments[2]);
+    let y_var = verifier.commit(witness_commitments[3]);
+
+    debug!("Applying dummy circuit constraints on verifier...");
+    apply_dummy_circuit_constraints(a_var, b_var, x_var, y_var, verifier);
+}
+
+pub fn get_dummy_proof() -> R1CSProof {
+    R1CSProof {
+        A_I1: StarkPoint::generator(),
+        A_O1: StarkPoint::generator(),
+        S1: StarkPoint::generator(),
+        A_I2: StarkPoint::generator(),
+        A_O2: StarkPoint::generator(),
+        S2: StarkPoint::generator(),
+        T_1: StarkPoint::generator(),
+        T_3: StarkPoint::generator(),
+        T_4: StarkPoint::generator(),
+        T_5: StarkPoint::generator(),
+        T_6: StarkPoint::generator(),
+        t_x: Scalar::random(&mut thread_rng()),
+        t_x_blinding: Scalar::random(&mut thread_rng()),
+        e_blinding: Scalar::random(&mut thread_rng()),
+        ipp_proof: InnerProductProof {
+            L_vec: vec![],
+            R_vec: vec![],
+            a: Scalar::random(&mut thread_rng()),
+            b: Scalar::random(&mut thread_rng()),
+        },
+    }
+}
+
+pub fn get_dummy_circuit_weights() -> CircuitWeights {
+    let mut transcript = HashChainTranscript::new(TRANSCRIPT_SEED.as_bytes());
+    let pc_gens = PedersenGens::default();
+    let mut prover = Prover::new(&pc_gens, &mut transcript);
+
+    let mut rng = thread_rng();
+
+    let (_, a_var) = prover.commit(Scalar::random(&mut rng), Scalar::random(&mut rng));
+    let (_, b_var) = prover.commit(Scalar::random(&mut rng), Scalar::random(&mut rng));
+    let (_, x_var) = prover.commit(Scalar::random(&mut rng), Scalar::random(&mut rng));
+    let (_, y_var) = prover.commit(Scalar::random(&mut rng), Scalar::random(&mut rng));
+
+    apply_dummy_circuit_constraints(a_var, b_var, x_var, y_var, &mut prover);
+
+    prover.get_weights()
+}
+
+pub fn get_dummy_circuit_params() -> CircuitParams {
+    let circuit_weights = get_dummy_circuit_weights();
+    let pc_gens = PedersenGens::default();
+    CircuitParams {
+        n: DUMMY_CIRCUIT_N,
+        n_plus: DUMMY_CIRCUIT_N_PLUS,
+        k: DUMMY_CIRCUIT_K,
+        q: DUMMY_CIRCUIT_Q,
+        m: DUMMY_CIRCUIT_M,
+        b: pc_gens.B,
+        b_blind: pc_gens.B_blinding,
+        w_l: circuit_weights.w_l,
+        w_o: circuit_weights.w_o,
+        w_r: circuit_weights.w_r,
+        w_v: circuit_weights.w_v,
+        c: circuit_weights.c,
     }
 }
